pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "gunicorn>=25.1.0",
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
]
